Make main.rs oblivious to the inner-workings of the emulator
Just export an Emulator (this was the Arch type in the Go version)
and call methods on that -> main doesn't have to know much at all,
except maybe argument parsing in the future.

This requires Chip8 itself to export the Chip8 struct
(with only one public function not in the Emulator trait) -
maybe there's a better way to do this?