Screen -> Hardware, implement some more instructions
Now everything except drawing sprites is completed
at a basic level. While we're here, we change a few
(& 0xFF) statements that aren't necessary because
of Rust's casting being C casts,
add the random number package as a dependency
and implement its instruction,
implement the basic loop of the emulation
(though it's not currently called)
and use slice immutable borrowing to avoid a loop.