Get tests working again and remove some dead code
Tests started breaking because now that Hardware::new()
creates an SDL object, the thread that calls Hardware::new()
must be the main thread. In our binary we can guarantee this,
but cargo does not guarantee this for tests (even with various options)
- it will always spawn at least one other thread and run code there,
which is completely fatal for testing.

A few approaches were evaluated to harmonize this, but the big
obstacle here now is serde, which requires Default to be implemented
for all types being deserialized. While we skip serializing/deserializing
the Hardware type, we cannot just define it as a trait object smart-pointer
(since Default on Box<> and similar seems to be not useful here).
So using a smart-pointer to a type that implements Interactible for example
doesn't quite work here. It's also possible to fix the situation by
making all the SDL-fields wrapped in Option<> (or even the Hardware member
in Chip8, for example), that requires a ton of code to deal with unpacking
Option<>s which is never going to fail in production, and would make it
harder to read the crate.

We settled on using the Rust version of #ifdef (alas, a C solution to
a Rust problem) - make the type of the Hardware object conditional in Chip8
based on whether or not we're running tests (tests get a different type,
MockHardware, that doesn't have SDL fields or anything else useful).
This also requires a few applications of the test check in module definitions,
etc...but this was by far the most compact way to implement this fix.
(That doesn't mean I'm entirely happy with it, etiher :) )

While we're here, to simplify using Hardware/MockHardware we move a few
functions to the Interactible trait, and we remove some commented out code
that was unnecessary (some manual serializing work / iterators for Screen,
which may come back later).