Style / magic number clean up
Remove 'return' in cases where the last line
yeilds the result without a semicolon.
Use a few constants and manually implement Default()
for Screen, which allows us to remove the manual specification
of screen resolution, which we weren't doing anyway.
In the future, screen resolution could also be provided by
the user at start-up time.

Fix the window title when a game state is loaded
and restore the unpause screen redraw.