Canonicalize import style and remove use of unwrap()
We standardize here on use of 'super::' when possible
and put local imports before std imports (which go before
external imports).

We add includes for a few common types like std::io::Error
and packages like std::fmt to slim the code just a bit.

Unwrap() we replace with expect() when we don't need to
format the message, otherwise we use unwrap_or_else()
with a closure.