Fix up a few small bugs and remove savestate debug code
Now setting debug on a loaded state works properly.
When loading a state that is not actually one (say, a regular game),
the error message printed on the CLI is now useful.

Also cleaned up a few magic numbers, some style, and removed
the debug printfs around the savestate code.